address PR comment
