+++ src/software/simulation/simulator_robot.h
@@ -3,6 +3,7 @@
 #include <cinttypes>
 #include <memory>
 
+#include "software/simulation/firmware_object_deleter.h"
 #include "software/simulation/physics/physics_ball.h"
 #include "software/simulation/physics/physics_robot.h"
 
@@ -11,23 +12,6 @@ extern "C"
 #include "firmware/app/primitives/primitive_manager.h"
 #include "shared/proto/primitive.nanopb.h"
 }
-
-/**
- * Because the PrimitiveManager_t struct is defined in the .c file (rather than the .h
- * file), C++ considers it an incomplete type and is unable to use it with smart pointers
- * because it doesn't know the size of the object. Therefore we need to create our own
- * "Deleter" class we can provide to the smart pointers to handle that instead.
- *
- * See https://en.cppreference.com/w/cpp/memory/unique_ptr/unique_ptr for more info and
- * examples
- */
-struct PrimitiveManagerDeleter
-{
-    void operator()(PrimitiveManager_t* primitive_manager) const
-    {
-        app_primitive_manager_destroy(primitive_manager);
-    };
-};
 
 /**
  * The SimulatorRobot class acts as a wrapper for a PhysicsRobot that deals with more
@@ -294,7 +278,7 @@ class SimulatorRobot
     // A pointer to all the balls currently being dribbled
     std::vector<PhysicsBall*> balls_in_dribbler_area;
 
-    std::unique_ptr<PrimitiveManager, PrimitiveManagerDeleter> primitive_manager;
+    std::unique_ptr<PrimitiveManager, FirmwarePrimitiveManagerDeleter> primitive_manager;
 
     // How much the dribbler damps the ball when they collide. Each component
     // of the damping can be changed separately so we have the flexibility to tune
